requirements.txt
requests>=2.29.0
Pillow>=9.0.0
RPi.GPIO>=0.7.0
spidev>=3.5 